moved next packet sequence query
